move higher-order function tools to localstack.util.functions
